+++ packages/core/types.js
@@ -1,181 +1,7 @@
-import { dirname, join as joinPath, relative, resolve } from "path";
-import {
-  getPackageConfig,
-  getParsedNpmDependencies,
-  getParsedNpmDevDependencies,
-} from "clio-manifest";
+import { cjs, clio as clioImport, esm, remote } from "./generator/imports.js";
 import { map, mapfn } from "bean-parser";
 
 import { SourceNode } from "source-map";
-import { existsSync } from "fs";
-
-class ImportError extends Error {
-  constructor(meta) {
-    super(meta.message);
-    this.meta = meta;
-  }
-}
-
-const ensureClioExtension = (path) =>
-  path.endsWith(".clio") ? path : path + ".clio";
-
-const toRelative = (path) => (path.match(/^\.{1,2}\//) ? path : "./" + path);
-
-const resolveRelativeModule = (context, path, line, column) => {
-  const currDir = dirname(
-    joinPath(context.srcPrefix, context.sourceDir, context.file)
-  );
-  const possiblePaths = [];
-  const resolvePath = resolve(currDir, path);
-
-  if (!resolvePath.endsWith(".clio")) {
-    possiblePaths.push(resolvePath + ".clio");
-  } else {
-    possiblePaths.push(resolvePath);
-  }
-  possiblePaths.push(joinPath(resolvePath, "main.clio"));
-
-  for (const path of possiblePaths) {
-    if (existsSync(path)) {
-      if (path.match(/^\.{1,2}\//)) {
-        return { source: path, require: path + ".js" };
-      }
-      const relativePath = relative(currDir, path);
-      return {
-        source: relativePath,
-        require: toRelative(relativePath + ".js"),
-      };
-    }
-  }
-
-  throw new ImportError({
-    message: [
-      `Cannot find module "${path}" in any of the following locations:\n`,
-      ...possiblePaths.map(
-        (path) => `  - ${relative(context.srcPrefix, path)}`
-      ),
-    ].join("\n"),
-    line,
-    column,
-  });
-};
-
-const resolveAbsoluteModule = (context, path, line, column) => {
-  const currDir = dirname(
-    joinPath(context.srcPrefix, context.sourceDir, context.file)
-  );
-  const possiblePaths = [];
-  const resolvePath = resolve(
-    context.srcPrefix,
-    context.sourceDir,
-    path.slice(1)
-  );
-
-  if (!resolvePath.endsWith(".clio")) {
-    possiblePaths.push(resolvePath + ".clio");
-  } else {
-    possiblePaths.push(resolvePath);
-  }
-  possiblePaths.push(joinPath(resolvePath, "main.clio"));
-
-  for (const path of possiblePaths) {
-    if (existsSync(path)) {
-      if (path.match(/^\.{1,2}\//)) {
-        return { source: path, require: path + ".js" };
-      }
-      const relativePath = relative(currDir, path);
-      return {
-        source: relativePath,
-        require: toRelative(relativePath + ".js"),
-      };
-    }
-  }
-
-  throw new ImportError({
-    message: [
-      `Cannot find module "${path}" in any of the following locations:\n`,
-      ...possiblePaths.map((path) => `  - ${relative(context.root, path)}`),
-    ].join("\n"),
-    line,
-    column,
-  });
-};
-
-const resolveModule = (context, path, line, column) => {
-  const [moduleName, subPath = ""] = path.match(/(.*?)(?:$|\/(.*))/).slice(1);
-  const modulePath = joinPath(context.modulesDir, moduleName);
-  if (!existsSync(modulePath)) {
-    throw new ImportError({
-      message: [
-        `Cannot find module "${moduleName}" in your project.`,
-        "Are you sure it is installed?",
-      ].join("\n"),
-      line,
-      column,
-    });
-  }
-
-  const { source, destination } = context.config.build;
-  const sourceDir = resolve(context.modulesDir, moduleName, source);
-  const possiblePaths = [];
-  const getResolvePath = (subPath) => {
-    return {
-      source: resolve(sourceDir, subPath),
-      destination: resolve(
-        context.modulesDestDir,
-        moduleName,
-        destination,
-        subPath
-      ),
-    };
-  };
-  if (!subPath.endsWith(".clio")) {
-    possiblePaths.push(getResolvePath(subPath + ".clio"));
-  } else {
-    possiblePaths.push(getResolvePath(subPath));
-  }
-  possiblePaths.push(getResolvePath(joinPath(subPath, "main.clio")));
-
-  for (const path of possiblePaths) {
-    if (existsSync(path.source)) {
-      if (path.destination.match(/^\.{1,2}\//)) {
-        return { source: path.source, require: path.destination + ".js" };
-      }
-      const requirePath = toRelative(
-        relative(dirname(context.destFile), path.destination) + ".js"
-      );
-      const sourcePath = relative(sourceDir, path.source);
-
-      return {
-        source: sourcePath,
-        require: requirePath,
-      };
-    }
-  }
-
-  throw new ImportError({
-    message: [
-      `Cannot find module "${path}" in any of the following locations:\n`,
-      ...possiblePaths.map((path) => `  - ${relative(context.root, path)}`),
-    ].join("\n"),
-    line,
-    column,
-  });
-};
-
-const getImportPath = (context, path, line, column) => {
-  if (!context.sourceDir) {
-    const source = ensureClioExtension(path);
-    return { source, require: source + ".js" };
-  }
-  if (path.match(/^\.{1,2}\//)) {
-    return resolveRelativeModule(context, path, line, column);
-  } else if (path.match(/^\//)) {
-    return resolveAbsoluteModule(context, path, line, column);
-  } else {
-    return resolveModule(context, path, line, column);
-  }
-};
 
 const join = (arr, sep) => new SourceNode(null, null, null, arr).join(sep);
 const asIs = (token) =>
@@ -183,7 +9,7 @@ const asIs = (token) =>
 
 const conditionals = ["fullConditional", "conditional"];
 
-const checkLambda = (node, body, context, getValue, getBody) => {
+export const checkLambda = (node, body, context, getValue, getBody) => {
   if (node.lambda?.length) {
     return get(
       {
@@ -384,85 +210,6 @@ const getTypeOf = (node, context) => {
   }
 };
 
-const compileImport = (path, importPath, context) => {
-  if (context.sourceDir) {
-    const isModule = !path.match(/^(\.{1,2})?\//);
-    const filePath = importPath.source;
-
-    const {
-      compileFile,
-      config,
-      configPath,
-      modulesDir,
-      modulesDestDir,
-      srcPrefix,
-      destPrefix,
-      cacheDir,
-      configs = {},
-      npmDeps = {},
-      npmDevDeps = {},
-      imports = [],
-    } = context;
-
-    if (isModule) {
-      // we need to get the config file
-      const moduleName = path.split("/").shift();
-      const configPath = joinPath(modulesDir, moduleName, "clio.toml");
-      const config = configs[moduleName] || getPackageConfig(configPath);
-      configs[moduleName] = config;
-      context.configs = configs;
-      const srcPrefix = joinPath(modulesDir, moduleName);
-      const destPrefix = joinPath(modulesDestDir, moduleName);
-      if (!npmDeps[moduleName]) {
-        npmDeps[moduleName] = getParsedNpmDependencies(configPath);
-        context.npmDeps = npmDeps;
-      }
-      if (!npmDevDeps[moduleName]) {
-        npmDevDeps[moduleName] = getParsedNpmDevDependencies(configPath);
-        context.npmDevDeps = npmDevDeps;
-      }
-      const { scope, destFile, srcFile } = compileFile(
-        filePath,
-        config,
-        configPath,
-        modulesDir,
-        modulesDestDir,
-        srcPrefix,
-        destPrefix,
-        cacheDir,
-        { npmDeps, npmDevDeps, configs }
-      );
-      imports.push({
-        module: moduleName,
-        src: srcFile,
-        dest: destFile,
-      });
-      context.imports = imports;
-      return scope;
-    } else {
-      const { scope, destFile, srcFile } = compileFile(
-        filePath,
-        config,
-        configPath,
-        modulesDir,
-        modulesDestDir,
-        srcPrefix,
-        destPrefix,
-        cacheDir,
-        { npmDeps, npmDevDeps, configs }
-      );
-      imports.push({
-        module: "",
-        src: srcFile,
-        dest: destFile,
-      });
-      context.imports = imports;
-      return scope;
-    }
-  }
-  return {};
-};
-
 const toProperCall = (node, context) => {
   const lastNode = node.content[node.content.length - 1];
   if (conditionals.includes(lastNode.type)) {
@@ -494,7 +241,7 @@ const toProperCall = (node, context) => {
   return node;
 };
 
-const types = {
+export const types = {
   symbol(node) {
     return asIs(node);
   },
@@ -1073,128 +820,18 @@ const types = {
         1. import
         2. assign
     */
-    const path = node.path.value.slice(1, -1).replace(/^[^:]*:/, "");
-    const filename = path.split("/").pop();
-    // Get the name, and make it pascalCase
-    const name = filename
-      .replace(/\.[^.]*$/, "")
-      .split("/")
-      .pop()
-      .replace(/@.*$/, "")
-      .split(/[-._]+/)
-      .filter(Boolean)
-      .map((v, i) => (i > 0 ? v[0].toUpperCase() + v.slice(1) : v))
-      .join("");
     const protocol =
       node.path.value.slice(1, -1).match(/^[^:]*(?=:)/)?.[0] || "clio";
-    let require;
-    let importScope = {};
-    /* We are clearly testing these in our unit tests,
-       yet istanbul considers them untested branches,
-       I have no idea why, I disabled istanbul checks
-       here until I figure out what is happening! */
-    if (protocol === "js") {
-      require = new SourceNode(
-        node.import.line,
-        node.import.column,
-        node.import.file,
-        ["require(", '"', path, '")']
-      );
-    } /* istanbul ignore next */ else if (protocol === "clio") {
-      /*
-        Resolve imports:
 
-          1. Check if import is relative or not
-          2. Check if import path exists
-          3. Check if import path is a directory or not
-          4. Check if we need to append .clio to the import path
-      
-      */
-
-      const { line, column } = node.path;
-      const importPath = getImportPath(context, path, line, column);
-
-      importScope = compileImport(path, importPath, context);
-
-      require = new SourceNode(
-        node.import.line,
-        node.import.column,
-        node.import.file,
-        ["await require(", '"', importPath.require, '").exports(clio)']
-      );
+    if (protocol === "cjs") {
+      return cjs(node, context, get);
+    } else if (protocol === "esm") {
+      return esm(node, context, get);
+    } else if (protocol === "clio") {
+      return clioImport(node, context, get);
     } else {
-      require = new SourceNode(
-        node.import.line,
-        node.import.column,
-        node.import.file,
-        ["await remote(clio,", '"', protocol + "://" + path, '")']
-      );
+      return remote(protocol, node, context, get);
     }
-    let assign;
-    if (!node.items) {
-      assign = new SourceNode(null, null, null, ["const ", name, "="]);
-      for (const key of Object.keys(importScope)) {
-        context.scope[`${name}.${key}`] = importScope[key];
-      }
-    } else {
-      let parts = [];
-      let rest;
-      for (const part of node.items) {
-        if (part.type === "symbol") {
-          const name = get(part, context);
-          parts.push(name);
-          if (importScope[name]) {
-            context.scope[name] = importScope[name];
-          }
-        } else if (part.lhs.type === "symbol") {
-          const name = get(part.lhs, context);
-          const rename = get(part.rhs, context);
-          parts.push(
-            new SourceNode(part.as.line, part.as.column, part.as.file, [
-              name,
-              ":",
-              rename,
-            ])
-          );
-          if (importScope[name]) {
-            context.scope[rename] = importScope[name];
-          }
-        } else {
-          const name = get(part.rhs, context);
-          rest = new SourceNode(part.as.line, part.as.column, part.as.file, [
-            "...",
-            name,
-          ]);
-          rest.name = name;
-          const importedNames = node.items
-            .filter(
-              (item) => item.type === "symbol" || item.lhs?.type === "symbol"
-            )
-            .map((item) => {
-              if (item.type === "symbol") {
-                return get(item, context).toString();
-              } else {
-                return get(item.rhs, context).toString();
-              }
-            });
-          for (const key of Object.keys(importScope)) {
-            if (!importedNames.includes(key)) {
-              context.scope[`${name}.${key}`] = importScope[key];
-            }
-          }
-        }
-      }
-      if (rest) parts.push(rest);
-      assign =
-        rest && parts.length == 1
-          ? new SourceNode(null, null, null, ["const ", rest.name, "="])
-          : new SourceNode(null, null, null, [
-              "const{",
-              new SourceNode(null, null, null, parts).join(","),
-              "}=",
-            ]);
-    }
-    return new SourceNode(null, null, null, [assign, require]);
   },
   lambda(node, context) {
     const { body } = node;
@@ -1764,33 +1401,29 @@ const types = {
     ]);
   },
   clio(node, context) {
-    const content = node.content
-      .map((node) => get(node, context))
+    const compiled = node.content.map((node) => get(node, context));
+    const content = compiled
       .map((node) => [node.insertBefore, node])
       .flat()
       .filter(Boolean);
-    const inner = new SourceNode(null, null, null, content).join(";");
+    const inner = new SourceNode(null, null, null, content);
+    const topLevels = compiled.map((part) => part.topLevel).filter(Boolean);
+    const outer = new SourceNode(null, null, null, topLevels);
     const builtins =
       "emitter,range,slice,remote,register,help,describe,returns,check,params,includes,f,Any";
     return new SourceNode(null, null, null, [
-      `module.exports.exports=async(clio)=>{const{${builtins}}=clio;`,
-      inner,
+      outer.join(";"),
+      `;export default async(clio)=>{const{${builtins}}=clio;`,
+      inner.join(";"),
       ";return clio.exports}",
     ]);
   },
 };
 
-const get = (node, context) => {
+export const get = (node, context) => {
   const result = types[node.type](node, context);
   result.node = node;
   return result;
 };
 
-const _ImportError = ImportError;
-export { _ImportError as ImportError };
-const _checkLambda = checkLambda;
-export { _checkLambda as checkLambda };
-const _types = types;
-export { _types as types };
-const _get = get;
-export { _get as get };
+export { ImportError } from "./generator/imports.js";
